spotR_server/requirements.txt
Flask
requests
python-dotenv
SQLAlchemy
Flask-Migrate
pytest
pytest-flask
